+++ core/src/main/java/org/elasticsearch/search/aggregations/metrics/percentiles/PercentilesAggregationBuilder.java
@@ -19,10 +19,14 @@
 
 package org.elasticsearch.search.aggregations.metrics.percentiles;
 
+import org.elasticsearch.common.ParseField;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.common.xcontent.ObjectParser;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.index.query.QueryParseContext;
 import org.elasticsearch.search.aggregations.AggregatorFactories.Builder;
+import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.aggregations.AggregatorFactory;
 import org.elasticsearch.search.aggregations.InternalAggregation.Type;
 import org.elasticsearch.search.aggregations.metrics.percentiles.hdr.HDRPercentilesAggregatorFactory;
@@ -34,6 +38,7 @@ import org.elasticsearch.search.aggregations.support.ValuesSource.Numeric;
 import org.elasticsearch.search.aggregations.support.ValuesSourceAggregationBuilder.LeafOnly;
 import org.elasticsearch.search.aggregations.support.ValuesSourceAggregatorFactory;
 import org.elasticsearch.search.aggregations.support.ValuesSourceConfig;
+import org.elasticsearch.search.aggregations.support.ValuesSourceParserHelper;
 import org.elasticsearch.search.aggregations.support.ValuesSourceType;
 
 import java.io.IOException;
@@ -44,7 +49,66 @@ public class PercentilesAggregationBuilder extends LeafOnly<ValuesSource.Numeric
     public static final String NAME = Percentiles.TYPE_NAME;
     public static final Type TYPE = new Type(NAME);
 
-    private double[] percents = PercentilesParser.DEFAULT_PERCENTS;
+    public static final double[] DEFAULT_PERCENTS = new double[] { 1, 5, 25, 50, 75, 95, 99 };
+    public static final ParseField PERCENTS_FIELD = new ParseField("percents");
+    public static final ParseField KEYED_FIELD = new ParseField("keyed");
+    public static final ParseField METHOD_FIELD = new ParseField("method");
+    public static final ParseField COMPRESSION_FIELD = new ParseField("compression");
+    public static final ParseField NUMBER_SIGNIFICANT_DIGITS_FIELD = new ParseField("number_of_significant_value_digits");
+
+    private static class TDigestOptions {
+        Double compression;
+    }
+
+    private static final ObjectParser<TDigestOptions, QueryParseContext> TDIGEST_OPTIONS_PARSER =
+            new ObjectParser<>(PercentilesMethod.TDIGEST.getParseField().getPreferredName(), TDigestOptions::new);
+    static {
+        TDIGEST_OPTIONS_PARSER.declareDouble((opts, compression) -> opts.compression = compression, COMPRESSION_FIELD);
+    }
+
+    private static class HDROptions {
+        Integer numberOfSigDigits;
+    }
+
+    private static final ObjectParser<HDROptions, QueryParseContext> HDR_OPTIONS_PARSER =
+            new ObjectParser<>(PercentilesMethod.HDR.getParseField().getPreferredName(), HDROptions::new);
+    static {
+        HDR_OPTIONS_PARSER.declareInt(
+                (opts, numberOfSigDigits) -> opts.numberOfSigDigits = numberOfSigDigits,
+                NUMBER_SIGNIFICANT_DIGITS_FIELD);
+    }
+
+    private static final ObjectParser<PercentilesAggregationBuilder, QueryParseContext> PARSER;
+    static {
+        PARSER = new ObjectParser<>(PercentilesAggregationBuilder.NAME);
+        ValuesSourceParserHelper.declareNumericFields(PARSER, true, true, false);
+
+        PARSER.declareDoubleArray(
+                (b, v) -> b.percentiles(v.stream().mapToDouble(Double::doubleValue).toArray()),
+                PERCENTS_FIELD);
+
+        PARSER.declareBoolean(PercentilesAggregationBuilder::keyed, KEYED_FIELD);
+
+        PARSER.declareField((b, v) -> {
+            b.method(PercentilesMethod.TDIGEST);
+            if (v.compression != null) {
+                b.compression(v.compression);
+            }
+        }, TDIGEST_OPTIONS_PARSER::parse, PercentilesMethod.TDIGEST.getParseField(), ObjectParser.ValueType.OBJECT);
+
+        PARSER.declareField((b, v) -> {
+            b.method(PercentilesMethod.HDR);
+            if (v.numberOfSigDigits != null) {
+                b.numberOfSignificantValueDigits(v.numberOfSigDigits);
+            }
+        }, HDR_OPTIONS_PARSER::parse, PercentilesMethod.HDR.getParseField(), ObjectParser.ValueType.OBJECT);
+    }
+
+    public static AggregationBuilder parse(String aggregationName, QueryParseContext context) throws IOException {
+        return PARSER.parse(context.parser(), new PercentilesAggregationBuilder(aggregationName), context);
+    }
+
+    private double[] percents = DEFAULT_PERCENTS;
     private PercentilesMethod method = PercentilesMethod.TDIGEST;
     private int numberOfSignificantValueDigits = 3;
     private double compression = 100.0;
@@ -180,13 +244,13 @@ public class PercentilesAggregationBuilder extends LeafOnly<ValuesSource.Numeric
 
     @Override
     protected XContentBuilder doXContentBody(XContentBuilder builder, Params params) throws IOException {
-        builder.array(PercentilesParser.PERCENTS_FIELD.getPreferredName(), percents);
-        builder.field(PercentilesParser.KEYED_FIELD.getPreferredName(), keyed);
+        builder.array(PERCENTS_FIELD.getPreferredName(), percents);
+        builder.field(KEYED_FIELD.getPreferredName(), keyed);
         builder.startObject(method.toString());
         if (method == PercentilesMethod.TDIGEST) {
-            builder.field(PercentilesParser.COMPRESSION_FIELD.getPreferredName(), compression);
+            builder.field(COMPRESSION_FIELD.getPreferredName(), compression);
         } else {
-            builder.field(PercentilesParser.NUMBER_SIGNIFICANT_DIGITS_FIELD.getPreferredName(), numberOfSignificantValueDigits);
+            builder.field(NUMBER_SIGNIFICANT_DIGITS_FIELD.getPreferredName(), numberOfSignificantValueDigits);
         }
         builder.endObject();
         return builder;
