+++ core/src/test/java/org/elasticsearch/search/aggregations/bucket/geogrid/GeoHashGridParserTests.java
@@ -39,9 +39,8 @@ public class GeoHashGridParserTests extends ESTestCase {
                 stParser, ParseFieldMatcher.STRICT);
         XContentParser.Token token = stParser.nextToken();
         assertSame(XContentParser.Token.START_OBJECT, token);
-        GeoHashGridParser parser = new GeoHashGridParser();
         // can create a factory
-        assertNotNull(parser.parse("geohash_grid", parseContext));
+        assertNotNull(GeoGridAggregationBuilder.parse("geohash_grid", parseContext));
     }
 
     public void testParseValidFromStrings() throws Exception {
@@ -51,9 +50,8 @@ public class GeoHashGridParserTests extends ESTestCase {
         QueryParseContext parseContext = new QueryParseContext(mockRegistry, stParser, ParseFieldMatcher.STRICT);
         XContentParser.Token token = stParser.nextToken();
         assertSame(XContentParser.Token.START_OBJECT, token);
-        GeoHashGridParser parser = new GeoHashGridParser();
         // can create a factory
-        assertNotNull(parser.parse("geohash_grid", parseContext));
+        assertNotNull(GeoGridAggregationBuilder.parse("geohash_grid", parseContext));
     }
 
     public void testParseErrorOnNonIntPrecision() throws Exception {
@@ -61,9 +59,8 @@ public class GeoHashGridParserTests extends ESTestCase {
         QueryParseContext parseContext = new QueryParseContext(mockRegistry, stParser, ParseFieldMatcher.STRICT);
         XContentParser.Token token = stParser.nextToken();
         assertSame(XContentParser.Token.START_OBJECT, token);
-        GeoHashGridParser parser = new GeoHashGridParser();
         try {
-            parser.parse("geohash_grid", parseContext);
+            GeoGridAggregationBuilder.parse("geohash_grid", parseContext);
             fail();
         } catch (ParsingException ex) {
             assertThat(ex.getCause(), instanceOf(NumberFormatException.class));
@@ -76,9 +73,8 @@ public class GeoHashGridParserTests extends ESTestCase {
         QueryParseContext parseContext = new QueryParseContext(mockRegistry, stParser, ParseFieldMatcher.STRICT);
         XContentParser.Token token = stParser.nextToken();
         assertSame(XContentParser.Token.START_OBJECT, token);
-        GeoHashGridParser parser = new GeoHashGridParser();
         try {
-            parser.parse("geohash_grid", parseContext);
+            GeoGridAggregationBuilder.parse("geohash_grid", parseContext);
             fail();
         } catch (IllegalArgumentException ex) {
             assertEquals("[geohash_grid] precision doesn't support values of type: VALUE_BOOLEAN", ex.getMessage());
@@ -90,9 +86,8 @@ public class GeoHashGridParserTests extends ESTestCase {
         QueryParseContext parseContext = new QueryParseContext(mockRegistry, stParser, ParseFieldMatcher.STRICT);
         XContentParser.Token token = stParser.nextToken();
         assertSame(XContentParser.Token.START_OBJECT, token);
-        GeoHashGridParser parser = new GeoHashGridParser();
         try {
-            parser.parse("geohash_grid", parseContext);
+            GeoGridAggregationBuilder.parse("geohash_grid", parseContext);
             fail();
         } catch (ParsingException ex) {
             assertThat(ex.getCause(), instanceOf(IllegalArgumentException.class));
